rename tensor_group_partition.py
