+++ easier/core/passes/__init__.py
@@ -5,7 +5,7 @@ from .metadata_propagation import propagate_metadata
 from .tensor_grouping import group_tensors
 from .reducer_binding import bind_reducer
 from .data_dependency_analysis import analyze_data_dependency
-from .tensor_partition import partition_tensor_groups
+from .tensor_group_partition import partition_tensor_groups
 from .sparse_encoding import encode_sparsity
 from .dataflow_distribution import distribute_dataflow
 # from .dataflow_fusion import fuse_dataflow
