+++ easier/core/passes/dataflow_distribution.py
@@ -23,7 +23,7 @@ from easier.core.runtime.modules import HaloExchanger, all_gather_into_tensor
 import easier.core.module as esr
 from easier.core.module import Module, Selector, Reducer
 from easier.core.passes.sparse_encoding.sparse_encoding import IdxMover
-from easier.core.passes.tensor_partition import \
+from easier.core.passes.tensor_group_partition import \
     ElemPart, partition_tensor_groups, insert_naive_elemparts
 from easier.core.passes.metadata_propagation.metadata import \
     EasierTensorMeta, Role, convert_scalar_type_to_torch_dtype, \
