+++ tests/core/passes/test_sparse_encoding.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock, Mock, patch
 
 import torch
 from easier.core.passes.tensor_grouping import EasierTensorGroup
-from easier.core.passes.tensor_partition import ElemPart as _EP_raw
+from easier.core.passes.tensor_group_partition import ElemPart as _EP_raw
 import easier.core.runtime.dist_env as _JitRuntimeDistEnv
 from easier.core.runtime.dist_env import DistEnv
 from easier.core.module import Selector, Reducer
